python/requirements.txt
asgiref==3.7.2
awscli==1.29.51
botocore==1.31.51
colorama==0.4.4
Django==4.1.11
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.4.2
docutils==0.16
jmespath==1.0.1
MarkupSafe==2.1.3
pyasn1==0.5.0
pymongo==3.12.3
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
rsa==4.7.2
s3transfer==0.6.2
six==1.16.0
sqlparse==0.2.4
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.7
Flask==1.0.3
Flask-Mail==0.9.1
Flask-WTF==0.14.2
zappa